# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shop Catalog API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: catalogs
  - name: products
  - name: users
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: Unauthorized

  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed

  /catalogs:
    get:
      operationId: listCatalogs
      tags: [catalogs]
      summary: List catalogs at root or within a parent
      parameters:
        - name: parentId
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        "200":
          description: List of catalogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Catalog"
    post:
      operationId: createCatalog
      tags: [catalogs]
      summary: Create a catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogBody"
      responses:
        "201":
          description: Created catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"

  /catalogs/{id}:
    get:
      operationId: getCatalog
      tags: [catalogs]
      summary: Get a catalog by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
    put:
      operationId: updateCatalog
      tags: [catalogs]
      summary: Update a catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogBody"
      responses:
        "200":
          description: Updated catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
    delete:
      operationId: deleteCatalog
      tags: [catalogs]
      summary: Delete a catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /catalogs/{id}/move:
    post:
      operationId: moveCatalog
      tags: [catalogs]
      summary: Move a catalog to another parent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveCatalogBody"
      responses:
        "200":
          description: Moved catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"

  /catalogs/{id}/breadcrumb:
    get:
      operationId: getCatalogBreadcrumb
      tags: [catalogs]
      summary: Get breadcrumb path for a catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Breadcrumb
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogBreadcrumbItem"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products in a catalog
      parameters:
        - name: catalogId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products/{id}/move:
    post:
      operationId: moveProduct
      tags: [products]
      summary: Move product to another catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveProductBody"
      responses:
        "200":
          description: Moved product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/bulk-delete:
    post:
      operationId: bulkDeleteProducts
      tags: [products]
      summary: Delete multiple products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteBody"
      responses:
        "204":
          description: Deleted

  /products/bulk-move:
    post:
      operationId: bulkMoveProducts
      tags: [products]
      summary: Move multiple products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMoveBody"
      responses:
        "204":
          description: Moved

  /attributes:
    get:
      operationId: listAttributes
      tags: [products]
      summary: List all custom attribute keys
      responses:
        "200":
          description: Attribute keys
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"

  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [user]

    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]

    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
      required: [id, username, role, createdAt]

    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
      required: [username, password, role]

    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]

    Catalog:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        childCount:
          type: integer
        productCount:
          type: integer
      required: [id, name, sortOrder, createdAt, childCount, productCount]

    CatalogBreadcrumbItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]

    CreateCatalogBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [name]

    UpdateCatalogBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer

    MoveCatalogBody:
      type: object
      properties:
        newParentId:
          type: integer
          nullable: true
      required: []

    ProductAttribute:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]

    Product:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        catalogId:
          type: integer
        imageUrl:
          type: string
          nullable: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, productId, name, price, catalogId, attributes, createdAt, updatedAt]

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        catalogId:
          type: integer
        imageUrl:
          type: string
          nullable: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        productId:
          type: string
      required: [name, price, catalogId]

    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        productId:
          type: string

    MoveProductBody:
      type: object
      properties:
        newCatalogId:
          type: integer
      required: [newCatalogId]

    BulkDeleteBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required: [ids]

    BulkMoveBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        newCatalogId:
          type: integer
      required: [ids, newCatalogId]